+++ src/decimal/udec/impls/from.rs
@@ -1,9 +1,31 @@
-use crate::decimal::{Decimal, ParseError, UnsignedDecimal};
+use crate::decimal::{Decimal, DecimalError, ParseError, UnsignedDecimal};
+
+type D<const N: usize> = Decimal<N>;
+type UD<const N: usize> = UnsignedDecimal<N>;
+
+impl<const N: usize> TryFrom<D<N>> for UD<N> {
+    type Error = DecimalError;
+
+    #[inline]
+    fn try_from(d: D<N>) -> Result<Self, Self::Error> {
+        if d.is_negative() {
+            return Err(DecimalError::Invalid);
+        }
+        Ok(Self::new(d))
+    }
+}
+
+impl<const N: usize> From<UD<N>> for D<N> {
+    #[inline]
+    fn from(d: UD<N>) -> Self {
+        d.0
+    }
+}
 
 macro_rules! from_uint {
     ($($uint: tt),*) => {
         $(
-            impl<const N: usize> From<$uint> for UnsignedDecimal<N> {
+            impl<const N: usize> From<$uint> for UD<N> {
                 #[inline]
                 fn from(int: $uint) -> Self {
                     Self::new(Decimal::from(int))
@@ -16,7 +38,7 @@ macro_rules! from_uint {
 macro_rules! try_from_int {
     ($($int: tt),*) => {
         $(
-            impl<const N: usize> TryFrom<$int> for UnsignedDecimal<N> {
+            impl<const N: usize> TryFrom<$int> for UD<N> {
                 type Error = ParseError;
 
                 #[inline]
@@ -34,7 +56,7 @@ macro_rules! try_from_int {
 from_uint!(u8, u16, u32, u64, u128, usize);
 try_from_int!(i8, i16, i32, i64, i128, isize);
 
-impl<const N: usize> TryFrom<f32> for UnsignedDecimal<N> {
+impl<const N: usize> TryFrom<f32> for UD<N> {
     type Error = ParseError;
 
     #[inline]
@@ -46,7 +68,7 @@ impl<const N: usize> TryFrom<f32> for UnsignedDecimal<N> {
     }
 }
 
-impl<const N: usize> TryFrom<f64> for UnsignedDecimal<N> {
+impl<const N: usize> TryFrom<f64> for UD<N> {
     type Error = ParseError;
 
     #[inline]